+++ SignNow.Net.Examples/ExamplesRunner.cs
@@ -157,6 +157,30 @@ namespace SignNow.Net.Examples
         }
 
         /// <summary>
+        /// Run test for example: <see cref="DocumentExamples.CreateTemplateFromTheDocument"/>
+        /// </summary>
+        [TestMethod]
+        [DataRow("Document Name")]
+        public async Task CreateDocumentFromTemplateTest(string documentName)
+        {
+            using var fileStream = File.OpenRead(PdfWithSignatureField);
+            var testDocumentId = (await testContext.Documents
+                .UploadDocumentWithFieldExtractAsync(fileStream, "CheckTheStatusOfTheDocument.pdf"))?.Id;
+            disposableDocumentId = testDocumentId;
+            var createTemplateRequest = new CreateTemplateFromDocumentRequest("TemplateName", testDocumentId);
+            var templateId = (await testContext.Documents.CreateTemplateFromDocumentAsync(createTemplateRequest)).Id;
+
+            var result = await DocumentExamples.CreateDocumentFromTheTemplate(templateId, documentName, token);
+            var document = await testContext.Documents.GetDocumentAsync(result.Id);
+
+            Assert.IsNotNull(document?.Id);
+            Assert.AreEqual(documentName, document.Name);
+
+            await testContext.Documents.DeleteDocumentAsync(document.Id);
+            await testContext.Documents.DeleteDocumentAsync(templateId);
+        }
+
+        /// <summary>
         /// Run test for example: <see cref="DocumentExamples.DownloadSignedDocument"/>
         /// </summary>
         [TestMethod]