+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - `ISignInvite.CancelEmbeddedInviteAsync` that allows to delete embedded signing invite for a document
 - `ISignInvite.GenerateEmbeddedInviteLinkAsync` that allows to create a link for the embedded signing invite for a document
 - `IDocumentService.CreateTemplateFromDocumentAsync` that allows to create template by flattening an existing document
+- `IDocumentService.CreateDocumentFromTemplateAsync` that allows to create document from the template
 
 ## Changed
 - Changed JsonConverter for `Model.SignNowInvite` properties